experimental/ner_wikiner_accelerated_with_speedster/requirements.txt
nebullvm>=0.4.4,<0.5.0
openvino-dev[onnx]
spacy-transformers>=1.1.0,<1.2.0
